requirements-dev.txt
pytest-tipsi-django==2.6.0
pytest==6.2.2
ipython==7.21.0

-r requirements.txt
